fix: Fix Order Book background color and add TradeHeader component
- Corrected background color styling in the Order Book component
- Added TradeHeader component to display trading section headings
