+++ apps/web/components/global/trade-token/index.tsx
@@ -5,7 +5,6 @@ import { getTradeData } from "@/lib/api/market-api";
 import Image from "next/image";
 import { tokenInfoType } from "../../../../../packages/lib/src/constants/tokenInfo";
 import { formatDate } from "@/utils/formatDate";
-import { Separator } from "@/components/ui/separator";
 import numeral from "numeral";
 
 type Props = {
@@ -28,28 +27,28 @@ const TradeTokenHeading = ({ tokenPair, type }: Props) => {
   console.log("type", type);
 
   return (
-    <div className="w-full bg-[#161616] text-nowrap">
-      {!data?.data ? (
-        <div>Data Not available</div>
-      ) : (
-        <div className="w-full flex gap-4 overflow-x-auto py-4 px-2 items-center">
-          <div className="flex gap-3 items-center min-w-[120px]">
-            <Image
-              src={data?.data?.icon}
-              alt={(data?.data?.token as string) || "icon"}
-              width={50}
-              height={50}
-              className="w-8 h-8"
-            />
-            <div className="flex flex-col gap-0">
-              <span className="text-[15px]">{data.data.token}</span>
-              <span className=" text-[12px] text-[#6A6A6A]">{formatDate()}</span>
+    <div className="bg-[#161616] w-full rounded-md">
+      <div className="text-nowrap w-full md:w-fit">
+        {!data?.data ? (
+          <div>Data Not available</div>
+        ) : (
+          <div className="grid grid-cols-2 md:grid-cols-6 py-3 px-4 md:px-2 items-center justify-end gap-2 md:gap-1">
+            <div className="flex gap-3 ">
+              <Image
+                src={data?.data?.icon}
+                alt={(data?.data?.token as string) || "icon"}
+                width={50}
+                height={50}
+                className="w-8 h-8 rounded-full"
+              />
+              <div className="flex flex-col gap-0">
+                <span className="text-[15px]">{data.data.token}</span>
+                <span className=" text-[12px] text-[#6A6A6A]">{formatDate()}</span>
+              </div>
+              {/* <Separator orientation="vertical" className="ml-2" /> */}
             </div>
-            <Separator orientation="vertical" className="ml-2" />
-          </div>
 
-          <div className="flex justify-center gap-6">
-            <div className="flex flex-col gap-0">
+            <div className="flex flex-col gap-0 md:ml-8 items-end md:items-start">
               <span className=" text-[14px] text-[#6A6A6A]">Price</span>
               <span className="text-[15px]">
                 {data.data.price?.toLocaleString("en-US", {
@@ -72,7 +71,7 @@ const TradeTokenHeading = ({ tokenPair, type }: Props) => {
               </span>
             </div>
 
-            <div className="flex flex-col gap-0">
+            <div className="flex flex-col gap-0 items-end md:items-start">
               <span className=" text-[14px] text-[#6A6A6A]">24h High</span>
               <span className="text-[15px]">
                 {data?.data?.high_24hr?.toLocaleString("en-US", {
@@ -92,13 +91,13 @@ const TradeTokenHeading = ({ tokenPair, type }: Props) => {
               </span>
             </div>
 
-            <div className="flex flex-col gap-0">
+            <div className="flex flex-col gap-0 items-end md:items-start">
               <span className=" text-[14px] text-[#6A6A6A]">Market Cap</span>
               <span className="text-[15px]">{numeral(data?.data?.market_cap).format("0.00a").toUpperCase()}</span>
             </div>
           </div>
-        </div>
-      )}
+        )}
+      </div>
     </div>
   );
 };
